backend/text-to-speech/requirements.txt
chatterbox-tts
fastapi
pydantic
modal
torchaudio